[ADD] product_supplierinfo_for_customer
Use product supplier info also for customers
============================================

This modules allows to use supplier info, available in _Procurements_ tab
from the product, also for customers, allowing to define prices per
customer and product.

For these prices to be used in sale prices calculations, you will have
to create a pricelist with a rule with option "Based on" is "Supplier prices
on the product form" (although the text is not clear enough).