[ADD] product_supplierinfo_discount_in_price
Discounts in product supplier info
==================================

This module includes discount field in product supplier info and apply it on
the price when you select in your pricelist that the prices are based on this
info.

NOTE: This module doesn't add a discount column on purchases orders, but blend
the result with it.